+++ example/rawkeys.dart
@@ -10,10 +10,18 @@ import 'package:dart_console/dart_console.dart';
 final console = Console();
 
 main() {
+  console.writeLine('Purely for testing purposes.');
+  console.writeLine();
+  console.writeLine(
+      'This method echos what stdin reads. Useful for testing unusual terminals.');
+  console.writeLine("Press 'q' to return to the command prompt.");
   console.rawMode = true;
 
   while (true) {
-    final codeUnit = stdin.readByteSync();
+    var codeUnit = 0;
+    while (codeUnit <= 0) {
+      codeUnit = stdin.readByteSync();
+    }
 
     if (codeUnit < 0x20 || codeUnit == 0x7F) {
       print('${codeUnit.toRadixString(16)}\r');