+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 0.2.0+2
+
+Bring back Dart 2.5 dependency, now that Dart 2.5 is released.
+
+## 0.2.0+1
+
+- Add escape character support
+
+## 0.2.0
+
+- Add `readline` function
+
+## 0.1.0+3
+
+- Tidy up documentation and comments
+
+## 0.1.0+2
+
+- Revert Dart version dependency, since it broke `dartdoc`
+
+## 0.1.0+1
+
+- Fix Dart version dependency
+
 ## 0.1.0
 
 - Initial version